# packages/shared-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vasooli API",
    "description": "AI-first revenue recovery — one policy layer, one audit trail, three engines.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/audit/batches/{batch_id}/summary": {
      "get": {
        "tags": [
          "audit"
        ],
        "summary": "Batch Summary",
        "description": "The headline metrics for one batch, recomputed from its entries.\n\nRecomputed rather than read from the batch's stored snapshot, so the number\nreported here and the trail behind it cannot disagree.",
        "operationId": "batch_summary_api_v1_audit_batches__batch_id__summary_get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/root-cause/runs": {
      "post": {
        "tags": [
          "root-cause"
        ],
        "summary": "Trigger Run",
        "description": "Run the full loop over a payments batch and return the honest summary.",
        "operationId": "trigger_run_api_v1_root_cause_runs_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__api__v1__routes__root_cause__RunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__engines__root_cause__schemas__RunSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "root-cause"
        ],
        "summary": "List Runs",
        "description": "Every root-cause run, newest first. Each carries its seed.",
        "operationId": "list_runs_api_v1_root_cause_runs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BatchRunRead"
                  },
                  "title": "Response List Runs Api V1 Root Cause Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/root-cause/runs/{batch_id}": {
      "get": {
        "tags": [
          "root-cause"
        ],
        "summary": "Get Run",
        "description": "One run's record, including the dataset batch id it consumed.",
        "operationId": "get_run_api_v1_root_cause_runs__batch_id__get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchRunRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/root-cause/runs/{batch_id}/summary": {
      "get": {
        "tags": [
          "root-cause"
        ],
        "summary": "Get Run Summary",
        "description": "The run's own summary — the engine-specific figures the dashboard reads.\n\nServed from what the run stored rather than recomputed, because the splits in\nit (per-determination diagnoses, detection scoring against the generator's\nground truth) need the run's outcomes and the dataset behind them, and nothing\nserving an API request should be re-running an engine.\n\nThe money figures inside it were read off the audit trail at the end of the\nrun, and `/audit/batches/{batch_id}/summary` recomputes those independently\n— so the stored copy is checkable against the trail rather than merely\nconvenient. Quote that route for a headline; quote this one for the breakdown.",
        "operationId": "get_run_summary_api_v1_root_cause_runs__batch_id__summary_get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__engines__root_cause__schemas__RunSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/root-cause/runs/{batch_id}/detections": {
      "get": {
        "tags": [
          "root-cause"
        ],
        "summary": "List Detections",
        "description": "Every corridor detection in a run, with its diagnosis and the rule that decided it.",
        "operationId": "list_detections_api_v1_root_cause_runs__batch_id__detections_get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          },
          {
            "name": "determination",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Determination"
            }
          },
          {
            "name": "allowed",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by whether the policy engine permitted the action.",
              "title": "Allowed"
            },
            "description": "Filter by whether the policy engine permitted the action."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CorridorDetectionRead"
                  },
                  "title": "Response List Detections Api V1 Root Cause Runs  Batch Id  Detections Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/root-cause/detections/{detection_id}": {
      "get": {
        "tags": [
          "root-cause"
        ],
        "summary": "Get Detection",
        "description": "One detection with its full diagnosis, reasoning and provenance.",
        "operationId": "get_detection_api_v1_root_cause_detections__detection_id__get",
        "parameters": [
          {
            "name": "detection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Detection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CorridorDetectionRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/root-cause/runs/{batch_id}/actions": {
      "get": {
        "tags": [
          "root-cause"
        ],
        "summary": "List Actions",
        "description": "Every action the run took or refused, each naming its authorising rule.",
        "operationId": "list_actions_api_v1_root_cause_runs__batch_id__actions_get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 200,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditEntryRead"
                  },
                  "title": "Response List Actions Api V1 Root Cause Runs  Batch Id  Actions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/root-cause/runs/{batch_id}/reroutes": {
      "get": {
        "tags": [
          "root-cause"
        ],
        "summary": "List Reroutes",
        "description": "Reroutes the run authorised, each with the moment it lapses.",
        "operationId": "list_reroutes_api_v1_root_cause_runs__batch_id__reroutes_get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          },
          {
            "name": "active_at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Return only reroutes in force at this instant. Expiry is not advisory.",
              "title": "Active At"
            },
            "description": "Return only reroutes in force at this instant. Expiry is not advisory."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CorridorRerouteRead"
                  },
                  "title": "Response List Reroutes Api V1 Root Cause Runs  Batch Id  Reroutes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchRunRead": {
        "properties": {
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "engine": {
            "$ref": "#/components/schemas/Engine"
          },
          "seed": {
            "type": "integer",
            "title": "Seed"
          },
          "status": {
            "$ref": "#/components/schemas/BatchStatus"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "notes": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "batch_id",
          "engine",
          "seed",
          "status",
          "started_at"
        ],
        "title": "BatchRunRead"
      },
...
      "BatchSummary": {
        "properties": {
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "entries": {
            "type": "integer",
            "title": "Entries"
          },
          "amount_at_risk_paise": {
            "type": "integer",
            "title": "Amount At Risk Paise"
          },
          "amount_recovered_paise": {
            "type": "integer",
            "title": "Amount Recovered Paise"
          },
          "recovery_rate": {
            "type": "number",
            "title": "Recovery Rate"
          },
          "action_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Action Counts"
          },
          "outcome_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Outcome Counts"
          },
          "engine_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Engine Counts"
          },
          "by_source": {
            "additionalProperties": {
              "$ref": "#/components/schemas/SourceBreakdown"
            },
            "type": "object",
            "title": "By Source"
          },
          "policy_violations": {
            "type": "integer",
            "title": "Policy Violations"
          },
          "blocked_count": {
            "type": "integer",
            "title": "Blocked Count"
          },
          "halted_count": {
            "type": "integer",
            "title": "Halted Count"
          },
          "escalated_count": {
            "type": "integer",
            "title": "Escalated Count"
          },
          "abstained_count": {
            "type": "integer",
            "title": "Abstained Count"
          }
        },
        "type": "object",
        "required": [
          "batch_id",
          "entries",
          "amount_at_risk_paise",
          "amount_recovered_paise",
          "recovery_rate",
          "action_counts",
          "outcome_counts",
          "engine_counts",
          "by_source",
          "policy_violations",
          "blocked_count",
          "halted_count",
          "escalated_count",
          "abstained_count"
        ],
        "title": "BatchSummary",
        "description": "The headline number, computed honestly from the log.\n\nNever from a separately maintained counter — a counter can drift from the\ntrail it claims to summarise, and then the demo's headline figure and its\nevidence disagree."
      },
...
      "CorridorRerouteRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "detection_id": {
            "type": "string",
            "title": "Detection Id"
          },
          "corridor_key": {
            "type": "string",
            "title": "Corridor Key"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "from_route_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Route Id"
          },
          "to_route_id": {
            "type": "string",
            "title": "To Route Id"
          },
          "effective_from": {
            "type": "string",
            "format": "date-time",
            "title": "Effective From"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "authorising_rule": {
            "type": "string",
            "title": "Authorising Rule"
          },
          "expiry_rule": {
            "type": "string",
            "title": "Expiry Rule"
          },
          "value_at_risk_paise": {
            "type": "integer",
            "title": "Value At Risk Paise"
          }
        },
        "type": "object",
        "required": [
          "id",
          "batch_id",
          "detection_id",
          "corridor_key",
          "method",
          "from_route_id",
          "to_route_id",
          "effective_from",
          "expires_at",
          "authorising_rule",
          "expiry_rule",
          "value_at_risk_paise"
        ],
        "title": "CorridorRerouteRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "app__api__v1__routes__root_cause__RunRequest": {
        "properties": {
          "batch_id": {
            "type": "string",
            "title": "Batch Id",
            "description": "Id for this engine run. Not the dataset's batch id."
          },
          "seed": {
            "type": "integer",
            "title": "Seed",
            "default": 42
          },
          "dataset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset",
            "description": "Path to a payments .jsonl. Defaults to the committed sample."
          },
          "now": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Now"
          },
          "score": {
            "type": "boolean",
            "title": "Score",
            "description": "Score detections against the generator's ground truth manifest.",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "batch_id"
        ],
        "title": "RunRequest",
        "description": "What to run, and against what.\n\n`now` is exposed because Phase 2's datasets are anchored to a fixed `as_of`\nthat is not today. Left unset, the runner derives it from the last attempt in\nthe dataset, which is the right default and the one the demo uses."
      },
...
      "app__engines__root_cause__schemas__RunSummary": {
        "properties": {
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "dataset_batch_id": {
            "type": "string",
            "title": "Dataset Batch Id"
          },
          "seed": {
            "type": "integer",
            "title": "Seed"
          },
          "now": {
            "type": "string",
            "format": "date-time",
            "title": "Now"
          },
          "attempts_ingested": {
            "type": "integer",
            "title": "Attempts Ingested"
          },
          "failed_attempts": {
            "type": "integer",
            "title": "Failed Attempts"
          },
          "detections": {
            "type": "integer",
            "title": "Detections"
          },
          "diagnoses_by_determination": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Diagnoses By Determination"
          },
          "amount_at_risk_paise": {
            "type": "integer",
            "title": "Amount At Risk Paise"
          },
          "amount_recovered_paise": {
            "type": "integer",
            "title": "Amount Recovered Paise"
          },
          "recovery_rate": {
            "type": "number",
            "title": "Recovery Rate"
          },
          "action_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Action Counts"
          },
          "outcome_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Outcome Counts"
          },
          "retries_suppressed": {
            "type": "integer",
            "title": "Retries Suppressed"
          },
          "retries_deferred": {
            "type": "integer",
            "title": "Retries Deferred"
          },
          "reroutes_authorised": {
            "type": "integer",
            "title": "Reroutes Authorised"
          },
          "policy_denials": {
            "type": "integer",
            "title": "Policy Denials"
          },
          "denials_by_rule": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Denials By Rule"
          },
          "human_escalations": {
            "type": "integer",
            "title": "Human Escalations"
          },
          "llm_fallbacks": {
            "type": "integer",
            "title": "Llm Fallbacks"
          },
          "provenance": {
            "additionalProperties": true,
            "type": "object",
            "title": "Provenance"
          },
          "detection_score": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DetectionScore"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "batch_id",
          "dataset_batch_id",
          "seed",
          "now",
          "attempts_ingested",
          "failed_attempts",
          "detections",
          "diagnoses_by_determination",
          "amount_at_risk_paise",
          "amount_recovered_paise",
          "recovery_rate",
          "action_counts",
          "outcome_counts",
          "retries_suppressed",
          "retries_deferred",
          "reroutes_authorised",
          "policy_denials",
          "denials_by_rule",
          "human_escalations",
          "llm_fallbacks",
          "provenance"
        ],
        "title": "RunSummary",
        "description": "The honest summary of one batch run.\n\nEvery money and count figure here is recomputed from the audit trail by\n`runner.py`, never accumulated in a side counter that could drift from it."
      }
    }
  }
}